+++ src/main/resources/tpch/queries.skew/14.sql
@@ -1,4 +1,4 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
@@ -12,5 +12,5 @@ from
 	part
 where
 	l_partkey = p_partkey
-	and l_shipdate >= date '1997-07-01'
-	and l_shipdate < date '1997-07-01' + interval '1' month
+	and l_shipdate >= date '1994-06-01'
+	and l_shipdate < date '1994-06-01' + interval '1' month
