+++ src/main/resources/tpch/queries.skew/18.sql
@@ -1,4 +1,4 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
@@ -18,9 +18,11 @@ where
 			l_orderkey
 		from
 			lineitem
+		where
+			l_quantity <= 100
 		group by
 			l_orderkey having
-				sum(l_quantity) > 313
+				sum(l_quantity) > 815
 	)
 	and c_custkey = o_custkey
 	and o_orderkey = l_orderkey
