+++ src/main/resources/tpch/queries.skew/19.sql
@@ -1,4 +1,4 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
@@ -9,9 +9,9 @@ from
 where
 	(
 		p_partkey = l_partkey
-		and p_brand = 'Brand#23'
+		and p_brand = 'Brand#00'
 		and p_container in ('SM CASE', 'SM BOX', 'SM PACK', 'SM PKG')
-		and l_quantity >= 8 and l_quantity <= 8 + 10
+		and l_quantity >= 7 and l_quantity <= 7 + 10
 		and p_size between 1 and 5
 		and l_shipmode in ('AIR', 'AIR REG')
 		and l_shipinstruct = 'DELIVER IN PERSON'
@@ -19,7 +19,7 @@ where
 	or
 	(
 		p_partkey = l_partkey
-		and p_brand = 'Brand#25'
+		and p_brand = 'Brand#00'
 		and p_container in ('MED BAG', 'MED BOX', 'MED PKG', 'MED PACK')
 		and l_quantity >= 16 and l_quantity <= 16 + 10
 		and p_size between 1 and 10
@@ -31,7 +31,7 @@ where
 		p_partkey = l_partkey
 		and p_brand = 'Brand#55'
 		and p_container in ('LG CASE', 'LG BOX', 'LG PACK', 'LG PKG')
-		and l_quantity >= 41 and l_quantity <= 41 + 10
+		and l_quantity >= 42 and l_quantity <= 42 + 10
 		and p_size between 1 and 15
 		and l_shipmode in ('AIR', 'AIR REG')
 		and l_shipinstruct = 'DELIVER IN PERSON'
