+++ src/main/resources/tpch/queries.skew/20.sql
@@ -1,4 +1,4 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
@@ -20,7 +20,7 @@ where
 				from
 					part
 				where
-					p_name like 'shiny m%'
+					p_name like 'shiny %'
 			)
 			and ps_availqty > (
 				select
@@ -30,11 +30,11 @@ where
 				where
 					l_partkey = ps_partkey
 					and l_suppkey = ps_suppkey
-					and l_shipdate >= date '1993-01-01'
-					and l_shipdate < date '1993-01-01' + interval '1' year
+					and l_shipdate >= date '1994-01-01'
+					and l_shipdate < date '1994-01-01' + interval '1' year
 			)
 	)
 	and s_nationkey = n_nationkey
-	and n_name = 'EGYPT'
+	and n_name = 'CHINA'
 order by
 	s_name
