+++ src/main/resources/tpch/queries.skew/22.sql
@@ -1,4 +1,4 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
@@ -14,7 +14,7 @@ from
 			customer
 		where
 			substring(c_phone, 1, 2) in
-				('40', '50', '60', '70', '80', '27', '16')
+				('40', '50', '60', '70', '80', '12', '18')
 			and c_acctbal > (
 				select
 					avg(c_acctbal)
@@ -23,7 +23,7 @@ from
 				where
 					c_acctbal > 0.00
 					and substring(c_phone, 1, 2) in
-						('40', '50', '60', '70', '80', '27', '16')
+						('40', '50', '60', '70', '80', '12', '18')
 			)
 			and not exists (
 				select
