+++ src/main/resources/tpch/queries.skew/4.sql
@@ -1,4 +1,4 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
@@ -7,8 +7,8 @@ select
 from
 	orders
 where
-	o_orderdate >= date '1997-04-01'
-	and o_orderdate < date '1997-04-01' + interval '3' month
+	o_orderdate >= date '1994-05-01'
+	and o_orderdate < date '1994-05-01' + interval '3' month
 	and exists (
 		select
 			*
