+++ src/main/resources/tpch/queries.skew/6.sql
@@ -1,4 +1,4 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
@@ -6,7 +6,7 @@ select
 from
 	lineitem
 where
-	l_shipdate >= date '1996-01-01'
-	and l_shipdate < date '1996-01-01' + interval '1' year
+	l_shipdate >= date '1993-01-01'
+	and l_shipdate < date '1993-01-01' + interval '1' year
 	and l_discount between 0.09 - 0.01 and 0.09 + 0.01
 	and l_quantity < 100
