+++ src/main/resources/tpch/queries.skew/8.sql
@@ -1,10 +1,10 @@
--- using 1498230177 as a seed to the RNG
+-- using 1499680242 as a seed to the RNG
 
 
 select
 	o_year,
 	sum(case
-		when nation = 'EGYPT' then volume
+		when nation = 'UNITED STATES' then volume
 		else 0
 	end) / sum(volume) as mkt_share
 from
@@ -29,7 +29,7 @@ from
 			and o_custkey = c_custkey
 			and c_nationkey = n1.n_nationkey
 			and n1.n_regionkey = r_regionkey
-			and r_name = 'MIDDLE EAST'
+			and r_name = 'AMERICA'
 			and s_nationkey = n2.n_nationkey
 			and o_orderdate between date '1995-01-01' and date '1996-12-31'
 			and p_type = 'SHINY MINED GOLD'
