+++ src/main/scala/com/databricks/spark/sql/perf/Benchmark.scala
@@ -283,7 +283,6 @@ case class Table(
     name: String,
     data: Dataset[_])
 
-
 object Benchmark {
 
   class ExperimentStatus(
