+++ src/main/scala/com/databricks/spark/sql/perf/tpcds/TPCDSTables.scala
@@ -1,77 +1,12 @@
-/*
- * Copyright 2015 Databricks Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.databricks.spark.sql.perf.tpcds
 
 import scala.sys.process._
 
-import org.slf4j.LoggerFactory
+import com.databricks.spark.sql.perf
+import com.databricks.spark.sql.perf.{DataGenerator, Table, Tables}
 
 import org.apache.spark.SparkContext
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.functions._
-import org.apache.spark.sql.types._
-import org.apache.spark.sql.{DataFrame, Row, SQLContext, SaveMode}
-
-trait DataGenerator extends Serializable {
-  def generate(
-    sparkContext: SparkContext,
-    name: String,
-    partitions: Int,
-    scaleFactor: String): RDD[String]
-}
-
-class DBGEN(dbgenDir: String, skew: Boolean = false) extends DataGenerator {
-  val dbgen = s"$dbgenDir/dbgen"
-  def generate(sparkContext: SparkContext,name: String, partitions: Int, scaleFactor: String) = {
-    val smallTables = Seq("nation", "region")
-    val numPartitions = if (partitions > 1 && !smallTables.contains(name)) partitions else 1
-    val generatedData = {
-      sparkContext.parallelize(1 to numPartitions, numPartitions).flatMap { i =>
-        val localToolsDir = if (new java.io.File(dbgen).exists) {
-          dbgenDir
-        } else if (new java.io.File(s"/$dbgenDir").exists) {
-          s"/$dbgenDir"
-        } else {
-          sys.error(s"Could not find dbgen at $dbgen or /$dbgenDir. Run install")
-        }
-        val parallel = if (numPartitions > 1) s"-C $partitions -S $i" else ""
-        val shortTableNames = Map(
-          "customer" -> "c",
-          "lineitem" -> "L",
-          "nation" -> "n",
-          "orders" -> "O",
-          "part" -> "P",
-          "region" -> "r",
-          "supplier" -> "s",
-          "partsupp" -> "S"
-        )
-        val skewParam = if (skew) "-k" else ""
-        val commands = Seq(
-          "bash", "-c",
-          s"cd $localToolsDir && ./dbgen -q $skewParam -T ${shortTableNames(name)} -s $scaleFactor $parallel")
-        println(commands)
-        commands.lines
-      }.repartition(numPartitions)
-    }
-
-    generatedData.setName(s"$name, sf=$scaleFactor, strings")
-    generatedData
-  }
-}
+import org.apache.spark.sql.SQLContext
 
 class DSDGEN(dsdgenDir: String) extends DataGenerator {
   val dsdgen = s"$dsdgenDir/dsdgen"
@@ -103,14 +38,13 @@ class DSDGEN(dsdgenDir: String) extends DataGenerator {
 }
 
 
-
 class TPCDSTables(
-    sqlContext: SQLContext,
-    dsdgenDir: String,
-    scaleFactor: String,
-    useDoubleForDecimal: Boolean = false,
-    useStringForDate: Boolean = false)
-    extends Tables(sqlContext, scaleFactor, useDoubleForDecimal, useStringForDate) {
+  sqlContext: SQLContext,
+  dsdgenDir: String,
+  scaleFactor: String,
+  useDoubleForDecimal: Boolean = false,
+  useStringForDate: Boolean = false)
+  extends Tables(sqlContext, scaleFactor, useDoubleForDecimal, useStringForDate) {
   import sqlContext.implicits._
 
   val dataGenerator = new DSDGEN(dsdgenDir)
@@ -589,348 +523,4 @@ class TPCDSTables(
       'web_gmt_offset           .decimal(5,2),
       'web_tax_percentage       .decimal(5,2))
   ).map(_.convertTypes())
-}
-
-class TPCHTables(
-  sqlContext: SQLContext,
-  dbgenDir: String,
-  skew: Boolean = false,
-  scaleFactor: String,
-  useDoubleForDecimal: Boolean = false,
-  useStringForDate: Boolean = false)
-  extends Tables(sqlContext, scaleFactor, useDoubleForDecimal, useStringForDate) {
-  import sqlContext.implicits._
-
-  val dataGenerator = new DBGEN(dbgenDir, skew)
-
-  val tables = Seq(
-    Table("part",
-      partitionColumns = Nil,
-      'p_partkey.long,
-      'p_name.string,
-      'p_mfgr.string,
-      'p_brand.string,
-      'p_type.string,
-      'p_size.int,
-      'p_container.string,
-      'p_retailprice.decimal,
-      'p_comment.string
-    ),
-    Table("supplier",
-      partitionColumns = Nil,
-      's_suppkey.long,
-      's_name.string,
-      's_address.string,
-      's_nationkey.long,
-      's_phone.string,
-      's_acctbal.decimal,
-      's_comment.string
-    ),
-    Table("partsupp",
-      partitionColumns = Nil,
-      'ps_partkey.long,
-      'ps_suppkey.long,
-      'ps_availqty.int,
-      'ps_supplycost.decimal,
-      'ps_comment.string
-    ),
-    Table("customer",
-      partitionColumns = Nil,
-      'c_custkey.long,
-      'c_name.string,
-      'c_address.string,
-      'c_nationkey.string,
-      'c_phone.string,
-      'c_acctbal.decimal,
-      'c_mktsegment.string,
-      'c_comment.string
-    ),
-    Table("orders",
-      partitionColumns = Nil,
-      'o_orderkey.long,
-      'o_custkey.long,
-      'o_orderstatus.string,
-      'o_totalprice.decimal,
-      'o_orderdate.date,
-      'o_orderpriority.string,
-      'o_clerk.string,
-      'o_shippriority.int,
-      'o_comment.string
-    ),
-    Table("lineitem",
-      partitionColumns = Nil,
-      'l_orderkey.long,
-      'l_partkey.long,
-      'l_suppkey.long,
-      'l_linenumber.int,
-      'l_quantity.decimal,
-      'l_extendedprice.decimal,
-      'l_discount.decimal,
-      'l_tax.decimal,
-      'l_returnflag.string,
-      'l_linestatus.string,
-      'l_shipdate.date,
-      'l_commitdate.date,
-      'l_receiptdate.date,
-      'l_shipinstruct.string,
-      'l_shipmode.string,
-      'l_comment.string
-    ),
-    Table("nation",
-      partitionColumns = Nil,
-      'n_nationkey.long,
-      'n_name.string,
-      'n_regionkey.long,
-      'n_comment.string
-    ),
-    Table("region",
-      partitionColumns = Nil,
-      'r_regionkey.long,
-      'r_name.string,
-      'r_comment.string
-    )
-  )
-}
-
-
-abstract class Tables(sqlContext: SQLContext, scaleFactor: String,
-    useDoubleForDecimal: Boolean = false, useStringForDate: Boolean = false)
-    extends Serializable {
-
-  def dataGenerator: DataGenerator
-  def tables: Seq[Table]
-
-  private val log = LoggerFactory.getLogger(getClass)
-
-  def sparkContext = sqlContext.sparkContext
-
-  case class Table(name: String, partitionColumns: Seq[String], fields: StructField*) {
-    val schema = StructType(fields)
-
-    def nonPartitioned: Table = {
-      Table(name, Nil, fields : _*)
-    }
-
-    /** 
-     *  If convertToSchema is true, the data from generator will be parsed into columns and
-     *  converted to `schema`. Otherwise, it just outputs the raw data (as a single STRING column).
-     */
-    def df(convertToSchema: Boolean, numPartition: Int) = {
-      val generatedData = dataGenerator.generate(sparkContext, name, numPartition, scaleFactor)
-      val rows = generatedData.mapPartitions { iter =>
-        iter.map { l =>
-          if (convertToSchema) {
-            val values = l.split("\\|", -1).dropRight(1).map { v =>
-              if (v.equals("")) {
-                // If the string value is an empty string, we turn it to a null
-                null
-              } else {
-                v
-              }
-            }
-            Row.fromSeq(values)
-          } else {
-            Row.fromSeq(Seq(l))
-          }
-        }
-      }
-
-      if (convertToSchema) {
-        val stringData =
-          sqlContext.createDataFrame(
-            rows,
-            StructType(schema.fields.map(f => StructField(f.name, StringType))))
-
-        val convertedData = {
-          val columns = schema.fields.map { f =>
-            col(f.name).cast(f.dataType).as(f.name)
-          }
-          stringData.select(columns: _*)
-        }
-
-        convertedData
-      } else {
-        sqlContext.createDataFrame(rows, StructType(Seq(StructField("value", StringType))))
-      }
-    }
-
-    def convertTypes(): Table = {
-      val newFields = fields.map { field =>
-        val newDataType = field.dataType match {
-          case decimal: DecimalType if useDoubleForDecimal => DoubleType
-          case date: DateType if useStringForDate => StringType
-          case other => other
-        }
-        field.copy(dataType = newDataType)
-      }
-
-      Table(name, partitionColumns, newFields:_*)
-    }
-
-    def genData(
-        location: String,
-        format: String,
-        overwrite: Boolean,
-        clusterByPartitionColumns: Boolean,
-        filterOutNullPartitionValues: Boolean,
-        numPartitions: Int): Unit = {
-      val mode = if (overwrite) SaveMode.Overwrite else SaveMode.Ignore
-
-      val data = df(format != "text", numPartitions)
-      val tempTableName = s"${name}_text"
-      data.registerTempTable(tempTableName)
-
-      val writer = if (partitionColumns.nonEmpty) {
-        if (clusterByPartitionColumns) {
-          val columnString = data.schema.fields.map { field =>
-            field.name
-          }.mkString(",")
-          val partitionColumnString = partitionColumns.mkString(",")
-          val predicates = if (filterOutNullPartitionValues) {
-            partitionColumns.map(col => s"$col IS NOT NULL").mkString("WHERE ", " AND ", "")
-          } else {
-            ""
-          }
-
-          val query =
-            s"""
-              |SELECT
-              |  $columnString
-              |FROM
-              |  $tempTableName
-              |$predicates
-              |DISTRIBUTE BY
-              |  $partitionColumnString
-            """.stripMargin
-          val grouped = sqlContext.sql(query)
-          println(s"Pre-clustering with partitioning columns with query $query.")
-          log.info(s"Pre-clustering with partitioning columns with query $query.")
-          grouped.write
-        } else {
-          data.write
-        }
-      } else {
-        data.write
-      }
-      writer.format(format).mode(mode)
-      if (partitionColumns.nonEmpty) {
-        writer.partitionBy(partitionColumns : _*)
-      }
-      println(s"Generating table $name in database to $location with save mode $mode.")
-      log.info(s"Generating table $name in database to $location with save mode $mode.")
-      writer.save(location)
-      sqlContext.dropTempTable(tempTableName)
-    }
-
-    def createExternalTable(location: String, format: String, databaseName: String,
-        overwrite: Boolean, discoverPartitions: Boolean = true): Unit = {
-
-      val qualifiedTableName = databaseName + "." + name
-      val tableExists = sqlContext.tableNames(databaseName).contains(name)
-      if (overwrite) {
-        sqlContext.sql(s"DROP TABLE IF EXISTS $databaseName.$name")
-      }
-      if (!tableExists || overwrite) {
-        println(s"Creating external table $name in database $databaseName using data stored in $location.")
-        log.info(s"Creating external table $name in database $databaseName using data stored in $location.")
-        sqlContext.createExternalTable(qualifiedTableName, location, format)
-      }
-      if (partitionColumns.nonEmpty && discoverPartitions) {
-        println(s"Discovering partitions for table $name.")
-        log.info(s"Discovering partitions for table $name.")
-        sqlContext.sql(s"ALTER TABLE $databaseName.$name RECOVER PARTITIONS")
-      }
-    }
-
-    def createTemporaryTable(location: String, format: String): Unit = {
-      println(s"Creating temporary table $name using data stored in $location.")
-      log.info(s"Creating temporary table $name using data stored in $location.")
-      sqlContext.read.format(format).load(location).registerTempTable(name)
-    }
-
-    def analyzeTable(databaseName: String, analyzeColumns: Boolean = false): Unit = {
-      println(s"Analyzing table $name.")
-      log.info(s"Analyzing table $name.")
-      sqlContext.sql(s"ANALYZE TABLE $databaseName.$name COMPUTE STATISTICS")
-      if (analyzeColumns) {
-        val allColumns = fields.map(_.name).mkString(", ")
-        println(s"Analyzing table $name columns $allColumns.")
-        log.info(s"Analyzing table $name columns $allColumns.")
-        sqlContext.sql(s"ANALYZE TABLE $databaseName.$name COMPUTE STATISTICS FOR COLUMNS $allColumns")
-      }
-    }
-  }
-
-  def genData(
-      location: String,
-      format: String,
-      overwrite: Boolean,
-      partitionTables: Boolean,
-      clusterByPartitionColumns: Boolean,
-      filterOutNullPartitionValues: Boolean,
-      tableFilter: String = "",
-      numPartitions: Int = 100): Unit = {
-    var tablesToBeGenerated = if (partitionTables) {
-      tables
-    } else {
-      tables.map(_.nonPartitioned)
-    }
-
-    if (!tableFilter.isEmpty) {
-      tablesToBeGenerated = tablesToBeGenerated.filter(_.name == tableFilter)
-      if (tablesToBeGenerated.isEmpty) {
-        throw new RuntimeException("Bad table name filter: " + tableFilter)
-      }
-    }
-
-    tablesToBeGenerated.foreach { table =>
-      val tableLocation = s"$location/${table.name}"
-      table.genData(tableLocation, format, overwrite, clusterByPartitionColumns,
-        filterOutNullPartitionValues, numPartitions)
-    }
-  }
-
-  def createExternalTables(location: String, format: String, databaseName: String,
-      overwrite: Boolean, discoverPartitions: Boolean, tableFilter: String = ""): Unit = {
-
-    val filtered = if (tableFilter.isEmpty) {
-      tables
-    } else {
-      tables.filter(_.name == tableFilter)
-    }
-
-    sqlContext.sql(s"CREATE DATABASE IF NOT EXISTS $databaseName")
-    filtered.foreach { table =>
-      val tableLocation = s"$location/${table.name}"
-      table.createExternalTable(tableLocation, format, databaseName, overwrite, discoverPartitions)
-    }
-    sqlContext.sql(s"USE $databaseName")
-    println(s"The current database has been set to $databaseName.")
-    log.info(s"The current database has been set to $databaseName.")
-  }
-
-  def createTemporaryTables(location: String, format: String, tableFilter: String = ""): Unit = {
-    val filtered = if (tableFilter.isEmpty) {
-      tables
-    } else {
-      tables.filter(_.name == tableFilter)
-    }
-    filtered.foreach { table =>
-      val tableLocation = s"$location/${table.name}"
-      table.createTemporaryTable(tableLocation, format)
-    }
-  }
-
-  def analyzeTables(databaseName: String, analyzeColumns: Boolean = false, tableFilter: String = ""): Unit = {
-    val filtered = if (tableFilter.isEmpty) {
-      tables
-    } else {
-      tables.filter(_.name == tableFilter)
-    }
-    filtered.foreach { table =>
-      table.analyzeTable(databaseName, analyzeColumns)
-    }
-  }
-
-
 }
